refactor(auth): remove jwt naming and align runtime tables
