+++ apps/api/src/auth/index.ts
@@ -1,7 +1,7 @@
 export * from './auth.module';
 export * from './services/auth.service';
 export * from './controllers/auth.controller';
-export * from './guards/jwt-auth.guard';
+export * from './guards/auth-session.guard';
 export * from './decorators/public.decorator';
 export * from './decorators/user.decorator';
 export * from './dto/auth.dto';
