+++ apps/api/src/directories/members.controller.ts
@@ -13,8 +13,8 @@ import {
 import { ApiTags, ApiBearerAuth, ApiOperation, ApiResponse } from '@nestjs/swagger';
 import { EventEmitter2 } from '@nestjs/event-emitter';
 import { DirectoryMemberService } from '@ever-works/agent/services';
-import { AuthService, CurrentUser, JwtAuthGuard } from '../auth';
-import { AuthenticatedUser } from '@src/auth/types/jwt.types';
+import { AuthService, CurrentUser, AuthSessionGuard } from '../auth';
+import { AuthenticatedUser } from '@src/auth/types/auth.types';
 import { InviteMemberDto } from './dto/invite-member.dto';
 import { UpdateMemberRoleDto } from './dto/update-member-role.dto';
 import { MemberInvitedEvent } from '../events';
@@ -25,7 +25,7 @@ import { ActivityActionType, ActivityStatus } from '@ever-works/agent/entities';
 @ApiTags('Members')
 @ApiBearerAuth('JWT-auth')
 @Controller('api/directories/:directoryId/members')
-@UseGuards(JwtAuthGuard)
+@UseGuards(AuthSessionGuard)
 export class MembersController {
     constructor(
         private readonly memberService: DirectoryMemberService,
